refactor(deck): share FREE_DECK_LIMIT, move useDeckGrace to composables/deck
Replace the duplicated deck-limit consts (subscription-actions, deck-grace) with
a single FREE_DECK_LIMIT in config/plans, and relocate useDeckGrace out of the
dashboard view dir to src/composables/deck/grace.ts so the upcoming locked-deck
editing + banner work can reuse it.
